+++ rastervision/data/raster_source/__init__.py
@@ -2,9 +2,7 @@
 
 from rastervision.data.raster_source.raster_source import *
 from rastervision.data.raster_source.raster_source_config import *
-from rastervision.data.raster_source.geotiff_source import *
-from rastervision.data.raster_source.geotiff_source_config import *
-from rastervision.data.raster_source.image_source import *
-from rastervision.data.raster_source.image_source_config import *
+from rastervision.data.raster_source.rasterio_source import *
+from rastervision.data.raster_source.rasterio_source_config import *
 from rastervision.data.raster_source.rasterized_source import *
 from rastervision.data.raster_source.rasterized_source_config import *