+++ rastervision/data/raster_source/default.py
@@ -8,39 +8,24 @@ class RasterSourceDefaultProvider(ABC):
     @staticmethod
     @abstractmethod
     def handles(s):
-        """Returns True of this provider is a default for this string"""
+        """Returns True if this provider is a default for this string"""
         pass
 
     @abstractmethod
     def construct(s, channel_order=None):
-        """Construts a default RasterSource based on the
-           string and an optional channel order
-        """
+        """Constructs default based on the string and an optional channel order."""
         pass
 
 
-class GeoTiffSourceDefaultProvider(RasterSourceDefaultProvider):
+class RasterioSourceDefaultProvider(RasterSourceDefaultProvider):
     @staticmethod
     def handles(uri):
         ext = os.path.splitext(uri)[1]
-        return ext.lower() in ['.tif', '.tiff']
+        return ext.lower() in ['.tif', '.tiff', '.geotiff', '.png', '.jpg']
 
     @staticmethod
     def construct(uri, channel_order=None):
-        return rv.RasterSourceConfig.builder(rv.GEOTIFF_SOURCE) \
-                                    .with_uri(uri) \
-                                    .with_channel_order(channel_order) \
-                                    .build()
-
-
-class ImageSourceDefaultProvider(RasterSourceDefaultProvider):
-    @staticmethod
-    def handles(uri):
-        return True  # This is the catch-all case.
-
-    @staticmethod
-    def construct(uri, channel_order=None):
-        return rv.RasterSourceConfig.builder(rv.IMAGE_SOURCE) \
+        return rv.RasterSourceConfig.builder(rv.RASTERIO_SOURCE) \
                                     .with_uri(uri) \
                                     .with_channel_order(channel_order) \
                                     .build()