+++ tests/command/test_eval_command.py
@@ -19,7 +19,7 @@ class TestEvalCommand(mk.MockMixin, unittest.TestCase):
             chip[:, :, :] *= np.array([0, 1, 2, 3]).astype(np.uint8)
             save_img(chip, img_path)
 
-            source = rv.data.ImageSourceConfig(img_path)
+            source = rv.data.RasterioSourceConfig(img_path)
 
             scenes = [rv.data.SceneConfig('scene_id', source)]
             evaluator = rv.EvaluatorConfig.builder(mk.MOCK_EVALUATOR).build()